workflows-multilang/python-runner/pyproject.toml
[project]
name = "multilang-tasking-runner-py"
version = "0.1.0"
description = "Tilebox Multilang Tasking Runner, Python"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "tilebox-workflows>=0.36.0",
]
